deinline fmpz_mod_mat and fmpz_mod_poly IO
