+++ src/fmpz_mod_mat.h
@@ -166,29 +166,14 @@ void fmpz_mod_mat_concat_vertical(fmpz_mod_mat_t res,
 }
 
 /* Input/output */
-FMPZ_MOD_MAT_INLINE
-int fmpz_mod_mat_fprint(FILE * file, const fmpz_mod_mat_t mat)
-{
-    return fmpz_mat_fprint(file, mat->mat);
-}
 
-FMPZ_MOD_MAT_INLINE
-int fmpz_mod_mat_fprint_pretty(FILE * file, const fmpz_mod_mat_t mat)
-{
-    return fmpz_mat_fprint_pretty(file, mat->mat);
-}
+#ifdef FLINT_HAVE_FILE
+int fmpz_mod_mat_fprint(FILE * file, const fmpz_mod_mat_t mat);
+int fmpz_mod_mat_fprint_pretty(FILE * file, const fmpz_mod_mat_t mat);
+#endif
 
-FMPZ_MOD_MAT_INLINE
-int fmpz_mod_mat_print(const fmpz_mod_mat_t mat)
-{
-    return fmpz_mat_print(mat->mat);
-}
-
-FMPZ_MOD_MAT_INLINE
-void fmpz_mod_mat_print_pretty(const fmpz_mod_mat_t mat)
-{
-    fmpz_mat_print_pretty(mat->mat);
-}
+int fmpz_mod_mat_print(const fmpz_mod_mat_t mat);
+void fmpz_mod_mat_print_pretty(const fmpz_mod_mat_t mat);
 
 /* Comparison */
 FMPZ_MOD_MAT_INLINE
